debugging

+++ model.py
@@ -283,7 +283,7 @@ class DeepVO(object):
         return tf.reduce_mean(tf.square(y_true - y_pred_fp))
 
     def sparse_categorical_crossentropy(self, ys, y_pred):
-        y_true = tf.clip_by_value(ys // self.bucket_size, 0, self.num_buckets - 1)
+        y_true = tf.to_int64(tf.clip_by_value(ys // self.bucket_size, 0, self.num_buckets - 1))
         print(y_true)
         cat_crossentropy_loss = tf.keras.losses.sparse_categorical_crossentropy(
             tf.one_hot(y_true, self.num_buckets), y_pred)
